Refactoring to move config to omniwitness
The omniwitness should not have been depending on the witness cmd internals. If we decide we don't want the witness cmd depending on the onmniwitness, then we need a third location for this to move to where we can have cmd and omniwitness point to that.
